requirements.txt
# Core orchestration
apache-airflow>=2.9.3,<3.0.0

# Web crawling & HTML parsing
requests>=2.32.0
beautifulsoup4>=4.12.3

# Configuration & utilities
pyyaml

# Rate limiting (important for polite & scalable crawling)
ratelimit>=2.2.1

# Better structured logging (highly recommended)
structlog>=24.4.0

# Optional but very useful additions (uncomment as needed)
# apache-airflow-providers-common-sql>=1.20.0     # if using SQL operators later
# boto3>=1.34.0                                 # for S3 storage instead of local files
# pandas>=2.2.2                                 # for more advanced metrics/aggregation
# tenacity>=9.0.0                               # for more robust retry logic

# Development / testing helpers 
pytest>=8.3.0
pytest-cov>=5.0.0